feat(studio): Add load to Multicall wrapper interface
